services/event-data-service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
SQLAlchemy==2.0.23
asyncpg==0.29.0
python-dotenv==1.0.1
httpx==0.27.0
aiofiles==23.2.1
strawberry-graphql[fastapi]==0.219.0
redis==5.0.1
aioredis==2.0.1
jinja2==3.1.2
python-multipart==0.0.6
